fix: fold decision envelope helper into agent.go for pkg file-count
